requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
requests==2.26.0
pandas==1.3.3
numpy==1.21.2
python-dotenv==0.19.1
pydantic==1.8.2
imutils==0.5.4
opencv-python==4.9.0.80